+++ BusinessObjects/Dtos/Auctions/AuctionListResponse.cs
@@ -9,7 +9,7 @@ public class AuctionListResponse
     public string Title { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
-    public int DepositFee { get; set; }
+    public decimal DepositFee { get; set; }
     public AuctionStatus Status { get; set; }
     public string ImageUrl { get; set; }
     public Guid ShopId { get; set; }
